+++ editor.php
@@ -222,6 +222,15 @@ if (!empty($_GET['nh'])) {
     }
 
     var connectEditor = function () {
+        if (typeof DocsAPI === 'undefined') {
+            var message = '<?php echo Display::return_message(
+                    $plugin->get_lang('OnlyofficeUnreachable'),
+                    'error'
+                )?>';
+            $("#cm-content .container").append(message);
+            return;
+        }
+
         $('link[rel="icon"]').attr(
             'href',
             '<?php echo Display::get_icon_path('../../plugin/onlyoffice/resources/' . $config['documentType'] .'.ico');?>'