+++ lib/onlyofficePlugin.php
@@ -34,7 +34,7 @@ class OnlyofficePlugin extends Plugin implements HookPluginInterface
     protected function __construct()
     {
         parent::__construct(
-            '1.5.0',
+            '1.6.0',
             'Asensio System SIA',
             [
                 'enable_onlyoffice_plugin' => 'boolean',