+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This plugin is developed and maintained at https://github.com/ONLYOFFICE/onlyoff
 - Catalan, Czech, Danish, Hungarian, Indonesian, Romanian, Albanian, Urdu empty file templates
 - support for office document question type
 - editor URL generation for documents with viewable extensions
+- plugin description and useful links in admin settings
 
 ## 1.5.0
 ## Added