+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This plugin is developed and maintained at https://github.com/ONLYOFFICE/onlyoff
 - support for office document question type
 - editor URL generation for documents with viewable extensions
 - plugin description and useful links in admin settings
+- hide a JWT secret from display
 
 ## 1.5.0
 ## Added